Illuminati/SSH: Support custom window/packet size values
Bumped net-ssh to patched version and enabled per-host custom packet
and window sizes to be set in the YAML configuration. The patched
version of net-ssh fixes any connectivity issues caused by trying to
open an SSH channel with a packet size larger than the size specified in
section 6.1 of RFC 4253.