Illuminati: Fix same-content check & Update bundle gems.
When writing to the tree, we hash the data to see if the object already
exists, however if the data is nothing, this does not work. Instead, we
skip committing entirely for anything empty. Also did a minor syntactic
cleanup on the null tree check (added parenthesis for clarity and
disposed of the pointless double tree setting)

Bundled gems have been updated to the latest revisions.