Illuminati: Handle channel open failure
Most hosts will not allow too many channels to be opened simultaneously,
thus, we wait 100ms if a failure occurs and then retry. We will retry
for a maximum of 20 times.

TODO: Detect if other threads are executing a channel successfully and
wait for one to close its channel if so (or allow user-specified channel
limits)