Illuminati: Code cleanup and simplification
This commit is the result of a large undertaking to cleanup the
underlying codebase without impacting the functionality of the program.

The primary aim has been to make the code simpler and easier to
understand, with the additional goal of also making minor improvements
in efficiency where possible. This commit will ultimately be merged into
master when all work is completed and the codebase is considered stable.