Illuminati/lib: Restore rescue of channel close and update Gemfile.
Previously we were handling the failure of a remote-side channel close
to ensure that we would still commit new data even if the remote side
closed the channel itself after reaching EOF. This change was lost
during the refactor and this commit restores it to prevent loss of host
SSH command data.

Also made a minor update to the Gemfile.lock and cleaned out a few
redundant case/when statements when handling SSH options.